# Use SSE2 in the ASCII fast path of decodeUtf8